移动一下方法
