+++ paicoding-ui/src/main/resources/templates/views/chat-home/sidebar/index.html
@@ -71,7 +71,7 @@
                         <span class="uno-buy-card-foot-price-unit">元</span>
                     </div>
                     <div class="uno-buy-card-foot-price-detail">
-                        <span class="uno-buy-card-foot-price-average">约0.24元/月</span>
+                        <span class="uno-buy-card-foot-price-average">约0.24元/天</span>
                         <span class="uno-buy-card-foot-price-original">原价119元</span>
                     </div>
                 </div>
