+++ paicoding-web/src/main/java/com/github/paicoding/forum/web/admin/rest/UserSettingRestController.java
@@ -1,8 +1,6 @@
 package com.github.paicoding.forum.web.admin.rest;
 
-import com.github.paicoding.forum.api.model.context.ReqInfoContext;
 import com.github.paicoding.forum.api.model.vo.ResVo;
-import com.github.paicoding.forum.api.model.vo.user.dto.BaseUserInfoDTO;
 import com.github.paicoding.forum.api.model.vo.user.dto.SimpleUserInfoDTO;
 import com.github.paicoding.forum.core.permission.Permission;
 import com.github.paicoding.forum.core.permission.UserRole;
@@ -41,12 +39,5 @@ public class UserSettingRestController {
         vo.setKey(key);
         vo.setItems(list);
         return ResVo.ok(vo);
-    }
-
-    @ApiOperation("获取当前登录用户信息")
-    @GetMapping("info")
-    public ResVo<BaseUserInfoDTO> info() {
-        BaseUserInfoDTO user = ReqInfoContext.getReqInfo().getUser();
-        return ResVo.ok(user);
     }
 }
